+++ scripts/aetherchrome.mjs
@@ -1,11 +1,22 @@
 import { AetherchromeActorData } from "./data/actor-data.mjs";
+import { AetherchromeItemData } from "./data/item-data.mjs";
 import { AetherchromeActorSheet } from "./sheets/actor-sheet.mjs";
+import { AetherchromeItemSheet } from "./sheets/item-sheet.mjs";
 
 Hooks.once("init", () => {
   console.log("Aetherchrome | Initializing");
 
   CONFIG.Actor.dataModels = {
     actor: AetherchromeActorData
+  };
+
+  CONFIG.Item.dataModels = {
+    weapon: AetherchromeItemData,
+    armor: AetherchromeItemData,
+    shield: AetherchromeItemData,
+    gear: AetherchromeItemData,
+    ammunition: AetherchromeItemData,
+    package: AetherchromeItemData
   };
 
   CONFIG.Actor.trackableAttributes = {
@@ -20,6 +31,12 @@ Hooks.once("init", () => {
     types: ["actor"],
     makeDefault: true,
     label: "AETHERCHROME.ActorSheet"
+  });
+
+  DocumentSheetConfig.registerSheet(Item, "aetherchrome", AetherchromeItemSheet, {
+    types: ["weapon", "armor", "shield", "gear", "ammunition", "package"],
+    makeDefault: true,
+    label: "AETHERCHROME.ItemSheet"
   });
 });